+++ src/const.ts
@@ -1,3 +1,3 @@
-export const CARD_VERSION = "1.4.3";
+export const CARD_VERSION = "1.4.4";
 export const CARD_TAG = "aerothermal-temperature-control";
 export const EDITOR_TAG = "aerothermal-temperature-control-editor";